refactor panelIdAlternativeMapping
